moved editForm to modal folder
